elec means electronic. I have chosen 6 musics for the background music. Currently 10 sound effects for power ups and no sound for enemies yet. The only problem is that the size of the wav files become very big when converted from mp3. So I included mp3 files too if you guys can convert them with smaller sizes.
For each level, a different music will be played. I'm not sure how many levels we are going to get, but I was thinking that electronic musics are used for levels that require speeds and jazz musics are used for levels that require more thinking and problem solving.

For giving credits, I just included a text file with links. I will find more sound effects for enemies and the player.